+++ run.py
@@ -1,6 +1,7 @@
 from library import app
+from library import socket_io
 
 this_app = app
 if __name__ == "__main__":
-    app.run(debug=True, use_debugger=True, use_reloader=False) 
-#     socket_io.run(app, port=os.environ.get("PORT"), log_output=True, debug=False)
+    # app.run(debug=True, use_debugger=True, use_reloader=False) 
+    socket_io.run(app)
